requirements.txt
Pillow
Django
django-filter
pyyaml
uritemplate
drf-yasg